+++ models/workDao.js
@@ -101,7 +101,7 @@ const worksList = async sort => {
 };
 
 // feed 상세
-const feed = async (id, user_id) => {
+const feed = async id => {
   try {
     // 조회수 카운팅 (IP주소나 시간만료 같은 장치는 아직 없음.)
     const viewCount = await myDataSource.query(
@@ -215,14 +215,6 @@ const feed = async (id, user_id) => {
         more_feed: JSON.parse(item.more_feed),
       };
     });
-    // feed 글쓴이와 유저와의 팔로우 관계
-    const checkFollow = await myDataSource.query(
-      `
-      select EXISTS (select f.id from Follow f
-        left join Works_Posting wp on wp.user_id = f.following_id
-        where wp.id = '${id}' and follower_id = '${user_id}') as success
-      `
-    );
 
     // feed 글쓴이에 대한 팔로워 정보
     let writerInfo = await myDataSource.query(
@@ -339,7 +331,6 @@ const feed = async (id, user_id) => {
       feedWithTags,
       feedCommentInfo,
       moreFeedinfo,
-      checkFollow,
       writerInfo,
       sympathyCount,
       sympathySortCount,
@@ -352,127 +343,7 @@ const feed = async (id, user_id) => {
   }
 };
 
-// -----------------------------------------------------------------------
-// follow 체결 관련
-const following = async (following_id, user_id) => {
-  try {
-    const follow = await myDataSource.query(
-      `
-    INSERT into Follow (following_id, follower_id) 
-    values ('${following_id}', '${user_id}') 
-    `
-    );
-    const followingResult = await myDataSource.query(
-      `
-    SELECT * from Follow f 
-    WHERE following_id = '${following_id}' and follower_id = '${user_id}'
-    `
-    );
-    let result = { followingResult };
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// follow 취소 관련
-const followingCancel = async (following_id, user_id) => {
-  try {
-    const deleteFollow = await myDataSource.query(
-      `
-    DELETE from Follow 
-    WHERE following_id = '${following_id}' and follower_id = '${user_id}'
-    `
-    );
-    const deleteResult = await myDataSource.query(
-      `
-    SELECT count(*) from Follow f 
-    WHERE following_id = '${following_id}' and follower_id = '${user_id}'
-    `
-    );
-    let result = { deleteResult };
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// 공감
-const sympathy = async (posting_id, user_id, sympathy_id) => {
-  try {
-    const checkSympathy = await myDataSource.query(
-      `
-    SELECT COUNT(*) check_cnt FROM Works_Sympathy_Count wsc
-    WHERE posting_id = '${posting_id}' and user_id = '${user_id}'
-    `
-    );
-    let checkValue = checkSympathy[0].check_cnt;
-    console.log('checkValue =', checkValue);
-    if (checkValue == 0) {
-      const insertSympathy = await myDataSource.query(
-        `
-      INSERT INTO Works_Sympathy_Count (user_id, posting_id, sympathy_id)
-      VALUES ('${user_id}', '${posting_id}', '${sympathy_id}')
-      `
-      );
-      const result = await myDataSource.query(
-        `
-      SELECT * FROM Works_Sympathy_Count wsc 
-      WHERE user_id = '${user_id}' and posting_id = '${posting_id}'
-      `
-      );
-      return result;
-    } else if (checkValue == 1) {
-      const insertSympathy = await myDataSource.query(
-        `
-      UPDATE Works_Sympathy_Count SET sympathy_id = '${sympathy_id}'
-      WHERE user_id = '${user_id}' and posting_id = '${posting_id}'
-      `
-      );
-      const result = await myDataSource.query(
-        `
-      SELECT * FROM Works_Sympathy_Count wsc 
-      WHERE user_id = '${user_id}' and posting_id = '${posting_id}'
-      `
-      );
-      return result;
-    }
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// 공감 취소
-const sympathyCancel = async (posting_id, user_id) => {
-  try {
-    const deleteSympathy = await myDataSource.query(
-      `
-    DELETE FROM Works_Sympathy_Count 
-    WHERE user_id = '${user_id}' and posting_id = '${posting_id}'
-    `
-    );
-    const checkSympathy = await myDataSource.query(
-      `
-    SELECT COUNT(*) check_cnt FROM Works_Sympathy_Count wsc
-    WHERE posting_id = '${posting_id}' and user_id = '${user_id}'
-    `
-    );
-    let result = { checkSympathy };
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
 module.exports = {
   worksList,
   feed,
-  following,
-  followingCancel,
-  sympathy,
-  sympathyCancel,
 };
