+++ routes/index.js
@@ -7,6 +7,8 @@ const categoryRouter = require('./categoryRouter');
 const feedRouter = require('./feedRouter');
 const workRouter = require('./workRouter');
 const searchListRouter = require('./searchListRouter');
+const sympathyRouter = require('./sympathyRouter');
+const followRouter = require('./followRouter');
 
 const uploadRouter = require('./uploadRouter');
 
@@ -16,4 +18,7 @@ router.use('/feeds', feedRouter);
 router.use('/category', categoryRouter);
 router.use('/searchlist', searchListRouter);
 router.use('/upload', uploadRouter);
+router.use('/sympathy', sympathyRouter);
+router.use('/follow', followRouter);
+
 module.exports = router;
