+++ services/workService.js
@@ -12,53 +12,9 @@ const worksList = async sort => {
 };
 
 // 지정된 피드 상세
-const feed = async (id, user_id) => {
+const feed = async id => {
   try {
-    const result = await workDao.feed(id, user_id);
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// 팔로우 체결
-const following = async (following_id, user_id) => {
-  try {
-    const result = await workDao.following(following_id, user_id);
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// 팔로우 체결
-const followingCancel = async (following_id, user_id) => {
-  try {
-    const result = await workDao.followingCancel(following_id, user_id);
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// 공감
-const sympathy = async (posting_id, user_id, sympathy_id) => {
-  try {
-    const result = await workDao.sympathy(posting_id, user_id, sympathy_id);
-    return result;
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
-  }
-};
-
-// 공감취소
-const sympathyCancel = async (posting_id, user_id) => {
-  try {
-    const result = await workDao.sympathyCancel(posting_id, user_id);
+    const result = await workDao.feed(id);
     return result;
   } catch (err) {
     console.log(err);
@@ -69,8 +25,4 @@ const sympathyCancel = async (posting_id, user_id) => {
 module.exports = {
   worksList,
   feed,
-  following,
-  followingCancel,
-  sympathy,
-  sympathyCancel,
 };
